refactor: clarify site server boundaries
