+++ scripts/activate_staged_deployment.sh
@@ -8,7 +8,7 @@ cd "$repo_root"
 service_name="${SERVICE_NAME:-okawak_blog}"
 service_file="${SERVICE_FILE:-service/okawak_blog.service}"
 systemd_unit_dir="${SYSTEMD_UNIT_DIR:-/etc/systemd/system}"
-target_bin="${TARGET_BIN:-./target/release/topcoat-server}"
+target_bin="${TARGET_BIN:-./target/release/site-server}"
 bin_dir="${BIN_DIR:-./bin}"
 staged_assets="${DEPLOY_STAGED_ASSETS:-./target/assets-staged}"
 live_assets="$bin_dir/assets"
